requirements.txt
apache-airflow-providers-postgres
requests==2.26.0
pandas==1.3.3
pyarrow==5.0.0
